Create Cusip.java
CUSIP Validator function (extra information):
A CUSIP (Committee on Uniform Securities Identification Procedures) is a nine-character alphanumeric code which act as a sort of DNA for the security, uniquely identifying a North American financial security or issuer along with the type of security for the purposes of facilitating clearing and settlement of trades. The CUSIP was adopted as an American National Standard under Accredited Standards X9.6. 

What each character represents is as follows:
    The first six characters identify the issuer and are assigned alphabetically.
    The seventh and eighth characters—which can be alphabetical or numerical—identify the type of security.
    The ninth and last digit is used as a check digit.

Example implementation of validating CUSIP Numbers:
Validate that the last digit (i.e., check digit) of the CUSIP code (1st column) is correct, against the following CUSIP Numbers:

      037833100       Apple Incorporated
      17275R102       Cisco Systems
      38259P508       Google Incorporated
      594918104       Microsoft Corporation
      68389X106       Oracle Corporation   (incorrect)
      68389X105       Oracle Corporation